--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -327,6 +327,7 @@ jobs:
           | grep [+-][0-9][0-9]\.[0-9][0-9]*[%] \
           | grep -v [-].* \
         )
+        echo "completed comparison logic"
         if [ -z "${nonNegativeComparisons}" ]; then
           echo "All max threshold comparisons are negative: this is acceptable"
         else