## Description

- Benchmark tests without plan cache.
- Benchmark tests statistics collection.
- Idiomatic `yaml` strings in doc-based views example.
- Diagnostics and thresholds.
- Comparison against historical benchmark baseline.
- Comparison against arbitrary benchmark maximum threshold.
- Becnhmark CI step enforces max threshold compliance.




<!-- Please provide a description of the change(s) implemented. -->

## Type of change

- [ ] Bug fix (non-breaking change to fix a bug).
- [x] Feature (non-breaking change to add functionality).
- [ ] Breaking change.
- [ ] Other (eg: documentation change).  **Please explain**.

## Issues referenced.

#314 

<!-- Please add deep links to any issues impacted by this PR. -->

## Evidence

From a previous run in a private repository, wherein the benchmarks were **accidentally** compared against historical baseline rather than hard limits, the below output proves that threshold overrun generates the desired CI failure:

```
##### Comparing to max threshold #####

pkg: github.com/stackql/stackql/internal/stackql/driver
                                                                │ cicd/ref/bench/baseline-go-bench.log │ cicd/log/current-bench-nocache.txt  │
                                                                │                sec/op                │    sec/op     vs base               │
SelectGoogleComputeInstanceDriver                                                         6.467m ± 19%   7.807m ± 17%  +20.71% (p=0.002 n=6)
ParallelProjectSelectGoogleComputeInstanceDriver                                          6.801m ±  1%   8.262m ±  2%  +21.47% (p=0.002 n=6)
HighlyParallelProjectSelectGoogleComputeInstanceDriver                                    11.90m ±  4%   14.15m ±  2%  +18.90% (p=0.002 n=6)
LoosenedHighlyParallelProjectSelectGoogleComputeInstanceDriver                            10.47m ±  2%   12.11m ±  2%  +15.68% (p=0.002 n=6)
UnlimitedHighlyParallelProjectSelectGoogleComputeInstanceDriver                           10.25m ±  2%   11.86m ±  2%  +15.69% (p=0.002 n=6)
geomean                                                                                   8.911m         10.56m        +18.47%
Some max threshold comparisons are positive or zero: this is unacceptable

##### Non-negative comparisons #####
+20.71%
+21.47%
+18.90%
+15.68%
+15.69%
+18.47%
```

<!-- Please add evidence (eg: test results, screen captures) that the changes are fit for purpose. -->

## Checklist:

- [x] A full round of testing has been completed, and there are no test failures as a result of these changes.
- [x] The changes are covered with functional and/or integration robot testing.
- [x] The changes work on all supported platforms.
- [x] Unit tests pass locally, as per [the developer guide](/docs/developer_guide.md#unit-tests).
- [x] Robot tests pass locally, as per [the developer guide](/docs/developer_guide.md#robot-tests).
- [x] Linter passes locally, as per [the developer guide](/docs/developer_guide.md#linting).

### Variations

N/A.

<!-- Please add fulsome explanations for any variations to the checklist. -->

## Tech Debt

No technical debt is introduced in this change.

<!-- If zero technical debt results from this change set, then please assert same here.  If, however, technical debt does result from this change set (the **strong** preference is that it does **not**), then please specify and justify.  Once assent is given in the PR conversation for the tech debt to be accrued, please include a link to an issue devoted to the tech debt, in the above issues section.  This should be done prior to merge. -->